Fixed data type mismatch in several registration classes (#77)
* Fixed data type mismatch in several registration classes

* chore: update changelog and bump version to 1.1.0 (#77)

* Bump version from 1.1.0 to 1.1.1


---------

+++ fireants/registration/greedy.py
@@ -83,7 +83,7 @@ class GreedyRegistration(AbstractRegistration, DeformableMixin):
         smooth_warp_sigma (float): Smoothing sigma for warp field
 
     """
-    def __init__(self, scales: List[int], iterations: List[float],
+    def __init__(self, scales: List[float], iterations: List[int],
                 fixed_images: BatchedImages, moving_images: BatchedImages,
                 loss_type: str = "cc",
                 deformation_type: str = 'compositive',
